// BassLines.Web/src/data/src/models/SongBaseWithImages.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BassLines
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    Image,
    ImageFromJSON,
    ImageFromJSONTyped,
    ImageToJSON,
} from './Image';

/**
 * 
 * @export
 * @interface SongBaseWithImages
 */
export interface SongBaseWithImages {
    /**
     * 
     * @type {string}
     * @memberof SongBaseWithImages
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof SongBaseWithImages
     */
    artist: string;
    /**
     * 
     * @type {string}
     * @memberof SongBaseWithImages
     */
    link?: string | null;
    /**
     * 
     * @type {Array<Image>}
     * @memberof SongBaseWithImages
     */
    images?: Array<Image> | null;
}

export function SongBaseWithImagesFromJSON(json: any): SongBaseWithImages {
    return SongBaseWithImagesFromJSONTyped(json, false);
}

export function SongBaseWithImagesFromJSONTyped(json: any, ignoreDiscriminator: boolean): SongBaseWithImages {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'title': json['title'],
        'artist': json['artist'],
        'link': !exists(json, 'link') ? undefined : json['link'],
        'images': !exists(json, 'images') ? undefined : (json['images'] === null ? null : (json['images'] as Array<any>).map(ImageFromJSON)),
    };
}

export function SongBaseWithImagesToJSON(value?: SongBaseWithImages | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'title': value.title,
        'artist': value.artist,
        'link': value.link,
        'images': value.images === undefined ? undefined : (value.images === null ? null : (value.images as Array<any>).map(ImageToJSON)),
    };
}